[eslint] rename conponent filename
